+++ database/schemas/tenant_settings_data.sql
@@ -1,16 +1,55 @@
-delete from public.tenant_settings cascade;
+DELETE FROM public.tenant_settings CASCADE;
 
-INSERT INTO public.tenant_settings (tenant_id, setting, description, value_type, value, created_at, created_by, updated_at, updated_by) VALUES
-(1, 'enable_inventory', 'use stock and COGS entry', 'BOOLEAN', 'TRUE', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system'),
-(1, 'enable_gst', 'Apply gst calculation', 'BOOLEAN', 'TRUE', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system'),
-(1, 'enable_bank_entry', 'Auto create bank receipt/payment', 'BOOLEAN', 'TRUE', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system'),
-(1, 'base_currency', 'Base currency for the tenant', 'CURRENCY', 'INR', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system')
-;
+-- Insert default settings for tenant 1
+INSERT INTO public.tenant_settings (
+    tenant_id, 
+    enable_inventory, 
+    enable_gst, 
+    enable_bank_entry, 
+    base_currency, 
+    payment_modes, 
+    default_payment_mode,
+    created_at, 
+    created_by, 
+    updated_at, 
+    updated_by
+) VALUES (
+    1, 
+    TRUE, 
+    TRUE, 
+    TRUE, 
+    'INR', 
+    ARRAY['CASH','UPI'], 
+    'CASH',
+    CURRENT_TIMESTAMP, 
+    'system', 
+    CURRENT_TIMESTAMP, 
+    'system'
+);
 
-
-INSERT INTO public.tenant_settings (tenant_id, setting, description, value_type, value, created_at, created_by, updated_at, updated_by) VALUES
-(2, 'enable_inventory', 'use stock and COGS entry', 'BOOLEAN', 'TRUE', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system'),
-(2, 'enable_gst', 'Apply gst calculation', 'BOOLEAN', 'TRUE', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system'),
-(2, 'enable_bank_entry', 'Auto create bank receipt/payment', 'BOOLEAN', 'TRUE', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system'),
-(2, 'base_currency', 'Base currency for the tenant', 'CURRENCY', 'INR', CURRENT_TIMESTAMP, 'system', CURRENT_TIMESTAMP, 'system')
-;
+-- Insert default settings for tenant 2
+INSERT INTO public.tenant_settings (
+    tenant_id, 
+    enable_inventory, 
+    enable_gst, 
+    enable_bank_entry, 
+    base_currency, 
+    payment_modes, 
+    default_payment_mode,
+    created_at, 
+    created_by, 
+    updated_at, 
+    updated_by
+) VALUES (
+    2, 
+    TRUE, 
+    TRUE, 
+    TRUE, 
+    'INR', 
+    ARRAY['CASH','UPI'], 
+    'CASH',
+    CURRENT_TIMESTAMP, 
+    'system', 
+    CURRENT_TIMESTAMP, 
+    'system'
+);